Split group member routes from core group routes
