+++ src/auth/mod.rs
@@ -1,4 +1,5 @@
 pub mod data;
 pub mod model;
+pub mod permissions;
 pub mod route;
 pub mod seed;
